relationship and owner Bootcamp and User

+++ controllers/bootcamps.js
@@ -34,6 +34,17 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @access   Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
 
+    // Add user to req.body
+    req.body.user = req.user.id;
+
+    // Check for published bootcamp
+    const publishedBootcamp = await Bootcamp.findOne({user: req.user.id});
+
+    // If the user is not an admin, they can only add one bootcamp
+    if(publishedBootcamp && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`The user with ${req.user.id} alredy published a bootcamp.`, 400))
+    }
+
     const bootcamp = await Bootcamp.create(req.body);
     res.status(201).json({
         sucess: true,
@@ -47,14 +58,21 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @access   Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-    });
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
     }
+
+    // Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.params.id} is not authorize to update this bootcamp`, 401));
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true
+    });
 
     res.status(200).json({
         sucess: true,
@@ -72,6 +90,11 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
+    }
+
+    // Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.params.id} is not authorize to delete this bootcamp`, 401));
     }
 
     bootcamp.deleteOne();
@@ -122,6 +145,11 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
+    }
+
+    // Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.params.id} is not authorize to update this bootcamp`, 401));
     }
